# Rewrite msvc backtrace support to be much faster on 64-bit platforms