@wesleywiser 2023-10-19
This will also fix https://github.com/rust-lang/rust/issues/116403 by virtue of not using `dbghelp.dll` during the initial backtrace capture.